local_requirements.txt
ai-models==0.3.1
aiodns==3.1.1
aiohttp==3.9.1
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
boto3==1.34.11
branca==0.7.0
Brotli==1.1.0
cdsapi==0.6.1
certifi==2023.11.17
cffi==1.16.0
cfgrib==0.9.10.4
cftime==1.6.3
charset-normalizer==3.3.2
click==8.1.7
climetlab==0.19.1
cloudpickle==3.0.0
colorama==0.4.6
cryptography==41.0.7
dask==2023.12.1
dm-tree==0.1.8
dnspython==2.4.2
earthkit-meteo==0.0.1
eccodes==1.6.1
ecmwf-api-client==1.6.3
ecmwf-opendata==0.2.0
ecmwflibs==0.5.3
email-validator==2.1.0.post1
entrypoints==0.4
exceptiongroup==1.2.0
fastapi==0.108.0
filelock==3.13.1
findlibs==0.0.5
frozenlist==1.4.1
fsspec==2023.12.2
GPUtil==1.4.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
imageio==2.33.1
importlib-metadata==7.0.1
inquirerpy==0.3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
locket==1.0.0
lru-dict==1.3.0
Magics==1.5.8
Markdown==3.5.1
MarkupSafe==2.1.3
multidict==6.0.4
multiurl==0.2.3.2
netCDF4==1.6.5
numpngw==0.1.3
numpy==1.26.2
orjson==3.9.10
packaging==23.2
pandas==2.1.4
paramiko==3.4.0
partd==1.4.1
pdbufr==0.11.0
pfzy==0.3.4
Pillow==10.1.0
prettytable==3.9.0
prompt-toolkit==3.0.43
py-cpuinfo==9.0.0
pycares==4.4.0
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.3.0
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyNaCl==1.5.0
pyodc==1.3.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
runpod==1.5.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.0
starlette==0.32.0.post1
termcolor==2.4.0
tomli==2.0.1
tomlkit==0.12.3
toolz==0.12.0
tqdm==4.66.1
tqdm-loggable==0.2
typing_extensions==4.9.0
tzdata==2023.4
ujson==5.9.0
urllib3==2.0.7
uvicorn==0.25.0
uvloop==0.19.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
xarray==2023.12.0
yarl==1.9.4
zipp==3.17.0
